Fix SSL verification
The current implementation of this library disables SSL verification, by
setting `verify_mode` to `OpenSSL::SSL::VERIFY_NONE` on the instantiated
`Net::HTTP` client.

This poses a potentially huge security risk for end-users, as it means
that the SSL certificate of Courier's servers is not confirmed to be
valid prior to sending the request. Subsequently, a malicious actor
could run a man-in-the-middle attack against folks using Courier and
intercept their requests, intercepting potentially sensitive user data
in the process.

This updates the `verify_mode` to be `OpenSSL::SSL::VERIFY_PEER`, which
correctly ensures that the SSL certificate of the server being connected
to is valid.

Further reading: https://brakemanscanner.org/docs/warning_types/ssl_verification_bypass/